[simulation_kernel] fix conversion warning
The cast is safe because we read only nBufferSize at most.

+++ src/libraries/simulation_kernel/Tools/MIL_Tools.cpp
@@ -66,7 +66,7 @@ boost::crc_32_type::value_type MIL_Tools::ComputeCRC( const std::string& fileNam
     while( file )
     {
         file.read( buffer, nBufferSize );
-        CRC.process_bytes( buffer, file.gcount() );
+        CRC.process_bytes( buffer, static_cast< size_t >( file.gcount() ));
     }
 
     file.close();
